(CDAP-16353) Move get artifact details for range into ArtifactRepositoryReader
Motivation:
AppLifecycleDeployment use getArtifactDetails to find the artifact detail
for deploymnet. Moving this method into ArtifactRepositoryReader allows us
to have remote vs local implication used by preview (when configured with
local levelDB) and appfabric respectively.

Add implementation for this method in both LocalArtifactRepositoryReader
and RemoteArtifactRepositoryReader. But the latter is not being used
anywhere at the moment. A follow-up change will make preview to use
RemoteArtifactRepositoryReader when configured with local levelDB.